Move some code to compat.py module
