+++ tests/compat.py
@@ -8,3 +8,22 @@ try:
     from urllib2 import urlopen  # noqa
 except ImportError:
     from urllib.request import urlopen  # noqa
+
+
+try:
+    from django.core.urlresolvers import is_valid_path
+except ImportError:
+    from django.core.urlresolvers import resolve, Resolver404
+
+    def is_valid_path(path, urlconf=None):
+        """Return True if path resolves against default URL resolver
+
+        This is a convenience method to make working with "is this a
+        match?" cases easier, avoiding unnecessarily indented
+        try...except blocks.
+        """
+        try:
+            resolve(path, urlconf)
+            return True
+        except Resolver404:
+            return False
