+++ tests/test_urls.py
@@ -1,24 +1,6 @@
 import pytest
 from django.conf import settings
-from .compat import force_text
-
-try:
-    from django.core.urlresolvers import is_valid_path
-except ImportError:
-    from django.core.urlresolvers import resolve, Resolver404
-
-    def is_valid_path(path, urlconf=None):
-        """Return True if path resolves against default URL resolver
-
-        This is a convenience method to make working with "is this a
-        match?" cases easier, avoiding unnecessarily indented
-        try...except blocks.
-        """
-        try:
-            resolve(path, urlconf)
-            return True
-        except Resolver404:
-            return False
+from .compat import force_text, is_valid_path
 
 
 @pytest.mark.urls('tests.urls_overridden')
